Moved migration generator to client_authenticator directory
