+++ lib/client_authenticator/version.rb
@@ -1,3 +1,3 @@
 module ClientAuthenticator
-  VERSION = "1.0.3"
+  VERSION = "1.0.4"
 end
