+++ lib/generators/client_authenticator/install_generator.rb
