+++ spec/spec_helper.rb
@@ -5,4 +5,4 @@ SimpleCov.minimum_coverage 88
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 require 'generator_spec'
 require 'client_authenticator'
-require 'generators/install_generator'
+require 'generators/client_authenticator/install_generator'
